requirements.txt
pymupdf
langchain
langchain-community
sentence-transformers
faiss-cpu
fastapi
uvicorn
streamlit
requests
python-multipart